Code review changes.

+++ db/migrate/20190510080621_add_failed_mfa_attempt_count_in_managers.rb
@@ -1,7 +1,7 @@
 class AddFailedMfaAttemptCountInManagers < DbMigrationConnection
   def up
     run_migration_for_db(DbConnection::KitClient) do
-      add_column :managers , :failed_mfa_attempt_count, :integer, after: :failed_login_attempt_count, null: true
+      add_column :managers , :failed_mfa_attempt_count, :tinyint, after: :failed_login_attempt_count, null: true
     end
   end
